+++ crates/adapters/src/integrated/delta_table/test.rs
@@ -24,7 +24,7 @@ use deltalake::protocol::SaveMode;
 use deltalake::table::config::TableProperty;
 use deltalake::{DeltaTable, DeltaTableBuilder, ensure_table_uri};
 use feldera_adapterlib::errors::controller::ControllerError;
-use feldera_macros::IsNone;
+use feldera_macros::{IsNone, OrdRepr};
 use feldera_sqllib::Variant;
 use feldera_types::config::{PipelineConfig, TransportConfig};
 use feldera_types::format::json::JsonFlavor;
@@ -4763,6 +4763,7 @@ async fn delta_table_follow_partition_column_types_test() {
     rkyv::Serialize,
     rkyv::Deserialize,
     IsNone,
+    OrdRepr,
 )]
 #[archive_attr(derive(Ord, Eq, PartialEq, PartialOrd))]
 struct UpperCaseTestStruct {
@@ -4839,6 +4840,7 @@ async fn delta_table_follow_partition_uppercase_columns_test() {
     rkyv::Serialize,
     rkyv::Deserialize,
     IsNone,
+    OrdRepr,
 )]
 #[archive_attr(derive(Ord, Eq, PartialEq, PartialOrd))]
 struct UnquotedSqlTestStruct {
@@ -6737,6 +6739,7 @@ async fn follow_filter_before_projection_prunes_scan() {
     rkyv::Serialize,
     rkyv::Deserialize,
     IsNone,
+    OrdRepr,
 )]
 #[archive_attr(derive(Ord, Eq, PartialEq, PartialOrd))]
 struct VariantTestStruct {