+++ crates/sqllib/src/variant.rs
@@ -8,7 +8,7 @@ use crate::{
 };
 use dbsp::algebra::{F32, F64};
 use feldera_fxp::DynamicDecimal;
-use feldera_macros::IsNone;
+use feldera_macros::{IsNone, OrdRepr};
 use feldera_types::serde_with_context::serde_config::VariantFormat;
 use feldera_types::serde_with_context::{
     DeserializeWithContext, SerializeWithContext, SqlSerdeConfig,
@@ -53,6 +53,7 @@ impl crate::__hash_repr::HashRepr for ArchivedVariant {
     rkyv::Serialize,
     rkyv::Deserialize,
     IsNone,
+    OrdRepr,
 )]
 #[archive(bound(
     serialize = "__S: rkyv::ser::ScratchSpace + rkyv::ser::Serializer + rkyv::ser::SharedSerializeRegistry",